ai_agents/sales_call_analyzer_v2/requirements.txt
# Core dependencies for Sales Call Analyzer V2
# LangGraph Agentic Framework with Gemini 2.5 Flash Lite

# LangChain and LangGraph (compatible versions)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# Google Generative AI
google-generativeai>=0.8.0

# Data processing
pandas>=2.2.0
numpy>=2.0.0

# Pydantic for schemas
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# LangSmith (optional but recommended for tracing)
langsmith>=0.2.0

# Type hints
typing-extensions>=4.0.0

